Creating a new wrapper struct for eigen methods
