+++ src/matrix/mod.rs
@@ -16,7 +16,7 @@ use error::{Error, ErrorKind};
 use utils;
 use vector::Vector;
 
-mod decomposition;
+pub mod decomposition;
 mod impl_ops;
 mod mat_mul;
 mod iter;
